Improve resource production calculation, and decouple basic income from being affected by production factor

+++ app/Models/Resources.php
@@ -54,10 +54,10 @@ class Resources
     /**
      * Multiply all resources by a factor.
      *
-     * @param int $factor
+     * @param float $factor
      * @return Resources
      */
-    public function multiply(int $factor): Resources
+    public function multiply(float $factor): Resources
     {
         $this->metal->multiply($factor);
         $this->crystal->multiply($factor);
